add permission control to some commands

+++ commands/moderation/prune.js
@@ -1,4 +1,4 @@
-const { SlashCommandBuilder } = require('discord.js');
+const { SlashCommandBuilder, PermissionsBitField } = require('discord.js');
 
 module.exports = {
 	data: new SlashCommandBuilder()
@@ -11,7 +11,17 @@ module.exports = {
 				.setMaxValue(100)),
 	async execute(interaction) {
 		const amount = interaction.options.getInteger('amount');
-
+		if (
+			!interaction.member.permissions.has(
+			  PermissionsBitField.Flags.ManageChannels
+			)
+		  )
+			//jika bukan moderator/admin (no manage channels)
+			return await interaction.reply({
+			  content: "You dont have permission to execute this command",
+			  ephemeral: true,
+			});
+	  
 		await interaction.channel.bulkDelete(amount, true).catch(error => {
 			console.error(error);
 			interaction.reply({ content: 'There was an error trying to prune messages in this channel!', ephemeral: true });
